+++ datamodel/openapi/openapi-api-sample/src/main/java/com/sap/cloud/sdk/datamodel/openapi/sample/model/OrderWithTimestamp.java
@@ -26,6 +26,7 @@ import java.util.Set;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
+import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -57,6 +58,7 @@ public class OrderWithTimestamp
     private OffsetDateTime timestamp;
 
     @JsonAnySetter
+    @JsonAnyGetter
     private final Map<String, Object> cloudSdkCustomFields = new LinkedHashMap<>();
 
     /**
@@ -304,6 +306,33 @@ public class OrderWithTimestamp
             throw new NoSuchElementException("OrderWithTimestamp has no field with name '" + name + "'.");
         }
         return cloudSdkCustomFields.get(name);
+    }
+
+    /**
+     * Put unrecognizable properties of the {@link OrderWithTimestamp}.
+     *
+     * @param customFields
+     *            The properties to be stored
+     */
+    @JsonIgnore
+    public void putAllCustomFields( @Nonnull Map<String, Object> customFields )
+    {
+        cloudSdkCustomFields.putAll(customFields);
+    }
+
+    /**
+     * Put an unrecognizable property of the {@link OrderWithTimestamp}. If the map previously contained a mapping for
+     * the key, the old value is replaced by the specified value.
+     *
+     * @param customFieldName
+     *            The name of the property
+     * @param customFieldValue
+     *            The value of the property
+     */
+    @JsonIgnore
+    public void putCustomField( @Nonnull String customFieldName, @Nonnull Object customFieldValue )
+    {
+        cloudSdkCustomFields.put(customFieldName, customFieldValue);
     }
 
     @Override