+++ datamodel/openapi/openapi-generator/src/test/resources/DataModelGeneratorIntegrationTest/api-class-vendor-extension-yaml/output/com/sap/cloud/sdk/services/apiclassvendorextension/model/NewSoda.java
@@ -29,6 +29,7 @@ import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.annotation.JsonValue;
 import java.io.Serializable;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
+import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.annotation.JsonTypeName;
@@ -57,6 +58,7 @@ public class NewSoda
   private Float price;
 
   @JsonAnySetter
+  @JsonAnyGetter
   private final Map<String, Object> cloudSdkCustomFields = new LinkedHashMap<>();
 
    /**
@@ -194,6 +196,27 @@ public class NewSoda
     }
     return cloudSdkCustomFields.get(name);
   }
+
+  /**
+   * Put unrecognizable properties of the {@link NewSoda}.
+   * @param customFields The properties to be stored
+   */
+  @JsonIgnore
+  public void putAllCustomFields( @Nonnull Map<String, Object> customFields ) {
+      cloudSdkCustomFields.putAll(customFields);
+  }
+
+  /**
+   * Put an unrecognizable property of the {@link NewSoda}.
+   * If the map previously contained a mapping for the key, the old value is replaced by the specified value.
+   * @param customFieldName The name of the property
+   * @param customFieldValue The value of the property
+   */
+  @JsonIgnore
+  public void putCustomField(@Nonnull String customFieldName, @Nonnull Object customFieldValue) {
+      cloudSdkCustomFields.put(customFieldName, customFieldValue);
+  }
+
 
   @Override
   public boolean equals(@Nullable final java.lang.Object o) {