+++ release_notes.md
@@ -32,7 +32,9 @@
 
 ### ✨ New Functionality
 
-- 
+- `OpenAPI` generated objects can now write custom fields which are not part of the object's schema:
+  - `putAllCustomFields(Map.of("nameOfField", "valueOfField", ...))`
+  - `putCustomField("nameOfField", "valueOfField")`
 
 ### 📈 Improvements